all exercises have an id field from the database
Former-commit-id: 693a2eb62658b40f2fc5d7f9f37e445e199437a3

+++ src/mitll/langtest/server/database/DAO.java
@@ -231,13 +231,13 @@ public class DAO {
     database.closeConnection(connection);
   }
 
-  protected long getGeneratedKey(PreparedStatement statement) throws SQLException {
+  protected int getGeneratedKey(PreparedStatement statement) throws SQLException {
     ResultSet rs = statement.getGeneratedKeys(); // will return the ID in ID_COLUMN
     long newID = -1;
     if (rs.next()) {
       newID = rs.getLong(1);
     }
-    return newID;
+    return (int)newID;
   }
 
   protected Connection getConnection() {
